Add a standalone build task
For CI/CD integration I needed a build task that does not watch files or launch a server. I decided to write this in JS instead of using npm tasks so that the steps of the build would be more composable.

I plan to port the local dev build to JS as well but wanted to get CI set up first.